+++ src/PhpSpreadsheet/Cell/Hyperlink.php
@@ -90,6 +90,14 @@ class Hyperlink
     }
 
     /**
+     * @return string
+     */
+    public function getTypeHyperlink()
+    {
+        return $this->isInternal() ? '' : 'External';
+    }
+
+    /**
      * Get hash code.
      *
      * @return string Hash code